move utility to internals
